Fixed pagination bug, locked self admin revoking.

+++ src/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -88,7 +92,10 @@ export class UsersService {
     return user;
   }
 
-  async deleteUser(userId: number) {
+  async deleteUser(userId: number, currentUser: User) {
+    if (userId === currentUser.id) {
+      throw new BadRequestException('WrongUserId');
+    }
     try {
       const result = await this.usersRepository.delete(userId);
       return result.affected > 0;
@@ -110,7 +117,11 @@ export class UsersService {
     return user;
   }
 
-  async revokeAdmin(userId: number) {
+  async revokeAdmin(userId: number, currentUser: User) {
+    if (userId === currentUser.id) {
+      throw new BadRequestException('WrongUserId');
+    }
+
     const user = await this.findById(userId);
     if (!user) {
       throw new NotFoundException(ErrorMessages.UserNotFound);
